Excell

+++ Core/AkinsoftExcellOkuma.Application/DTOs/DayDTO.cs
@@ -9,7 +9,7 @@ namespace AkinsoftExcellOkuma.Application.DTOs
     public class DayDTO
     {
         public string SicilNo { get; set; }
-        public int AyKodu { get; set; }
+        public short AyKodu { get; set; }
         public string IsyeriKodu { get; set; }
         public string IsyeriAdi { get; set; }
         public int CL { get; set; }
@@ -21,14 +21,14 @@ namespace AkinsoftExcellOkuma.Application.DTOs
         public int YL { get; set; }
         public int GM { get; set; }
         public string EksikGun { get; set; }
-        public int ToplamPrimGun { get; set; }
-        public int ToplamUcretGun { get; set; }
+        public short ToplamPrimGun { get; set; }
+        public short ToplamUcretGun { get; set; }
         public double MESAI_HFICI_PNC { get; set; }
         public double MESAI_HFSN_PNC { get; set; }
         public double MESAI_BYRM_PNC { get; set; }
         public double MESAI_HFICI_GECE_PNC { get; set; }
         public double MESAI_HFSN_GECE_PNC { get; set; }
         public double MESAI_BYRM_GECE_PNC { get; set; }
-        public int TOPLAM_UZAKTAN_CALISMA { get; set; }
+        public short TOPLAM_UZAKTAN_CALISMA { get; set; }
     }
 }
